fix: address PR review — layering, validation, and error handling
- Remove LIST preflight for channel read/message — channels created on
  JOIN won't be in LIST, and partial LIST results cause false negatives
- Catch TimeoutError (empty message) in channel dispatch error handler
- Move relative_time() to culture/formatting.py (non-CLI module) to
  avoid cross-layer dependency from observer/overview into cli/shared
- Relax server default validation: accept configured server name and
  servers with PID files (not just currently running)
